Initial and final commit
So this project was done a year ago (almost) when I wasn't using git, and because of that, I can commit only one time. For future projects that will require the use of git, I will commit multiple times, to see a steady progress in the development of the project.